requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
joblib==1.3.2
scikit-learn==1.3.0
pandas==2.2.2
python-dotenv==1.0.0
